refactor(backend): introduce connector feature package
Moves the connector/signal controllers, webhook dispatch/validation,
connector scopes, entities, repositories and DTOs out of the generic
controller/events/security/database/dto layer packages into a new
net.onelitefeather.blackhole.backend.connector package. The generic
domain-event bus (DomainEvent/DomainEventPublisher/RabbitTopology/
BlackholeRabbitTopology) stays in events/ as genuinely cross-domain
infra; SecretHasher stays in utils/ since it also serves elo's
ChatToxicityService, not just connector registration.

Removes the now-empty controller/dto/database.entities/database.
repository/cache/redis/response/security/scheduling layer packages,
completing the package-by-layer to package-by-feature migration
across all 8 backend domains. Pure physical move, no behavior change.
