+++ backend/src/main/java/net/onelitefeather/blackhole/backend/connector/ConnectorRegistrationEntity.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.database.entities;
+package net.onelitefeather.blackhole.backend.connector;
 
 import io.micronaut.serde.annotation.Serdeable;
 import jakarta.persistence.CollectionTable;
@@ -14,8 +14,6 @@ import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.Table;
 import net.onelitefeather.blackhole.backend.database.converter.MapStringObjectConverter;
-import net.onelitefeather.blackhole.backend.dto.ConnectorRegistrationDTO;
-import net.onelitefeather.blackhole.backend.dto.ConnectorStatus;
 import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
