+++ backend/src/main/java/net/onelitefeather/blackhole/backend/connector/ConnectorRegistrationRepository.java
@@ -1,8 +1,7 @@
-package net.onelitefeather.blackhole.backend.database.repository;
+package net.onelitefeather.blackhole.backend.connector;
 
 import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.repository.PageableRepository;
-import net.onelitefeather.blackhole.backend.database.entities.ConnectorRegistrationEntity;
 
 import java.util.Optional;
 import java.util.UUID;
