+++ backend/src/main/java/net/onelitefeather/blackhole/backend/connector/ConnectorRegistrationRequestDTO.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.dto;
+package net.onelitefeather.blackhole.backend.connector;
 
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.annotation.ReflectiveAccess;
@@ -7,13 +7,11 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 
 import java.util.List;
-import java.util.UUID;
 
 @Serdeable
 @ReflectiveAccess
-public record EventSubscriptionRequestDTO(
-        @NonNull UUID connectorId,
-        @NonNull @NotEmpty List<String> eventTypes,
-        @NonNull @NotBlank String deliveryUrl
+public record ConnectorRegistrationRequestDTO(
+        @NonNull @NotBlank String name,
+        @NonNull @NotEmpty List<String> scopes
 ) {
 }
