+++ backend/src/main/java/net/onelitefeather/blackhole/backend/connector/ConnectorStatus.java
@@ -1,4 +1,4 @@
-package net.onelitefeather.blackhole.backend.dto;
+package net.onelitefeather.blackhole.backend.connector;
 
 /**
  * Whether a registered connector may currently authenticate/receive webhook deliveries.
